Refactor a large portion of the payments code to prep it for PayPal support.
