+++ app/src/main/java/org/thoughtcrime/securesms/badges/gifts/flow/GiftFlowActivity.kt
@@ -11,14 +11,14 @@ import io.reactivex.rxjava3.subjects.Subject
 import org.thoughtcrime.securesms.R
 import org.thoughtcrime.securesms.components.FragmentWrapperActivity
 import org.thoughtcrime.securesms.components.settings.app.subscription.DonationPaymentComponent
-import org.thoughtcrime.securesms.components.settings.app.subscription.DonationPaymentRepository
+import org.thoughtcrime.securesms.components.settings.app.subscription.StripeRepository
 
 /**
  * Activity which houses the gift flow.
  */
 class GiftFlowActivity : FragmentWrapperActivity(), DonationPaymentComponent {
 
-  override val donationPaymentRepository: DonationPaymentRepository by lazy { DonationPaymentRepository(this) }
+  override val stripeRepository: StripeRepository by lazy { StripeRepository(this) }
 
   override val googlePayResultPublisher: Subject<DonationPaymentComponent.GooglePayResult> = PublishSubject.create()
