+++ app/src/main/java/org/thoughtcrime/securesms/components/settings/app/subscription/DonationPaymentComponent.kt
@@ -6,7 +6,7 @@ import io.reactivex.rxjava3.subjects.Subject
 import kotlinx.parcelize.Parcelize
 
 interface DonationPaymentComponent {
-  val donationPaymentRepository: DonationPaymentRepository
+  val stripeRepository: StripeRepository
   val googlePayResultPublisher: Subject<GooglePayResult>
 
   @Parcelize
