--- app/src/main/java/org/thoughtcrime/securesms/components/settings/app/subscription/SubscriptionsRepository.kt
@@ -1,68 +1,0 @@
-package org.thoughtcrime.securesms.components.settings.app.subscription
-
-import io.reactivex.rxjava3.core.Completable
-import io.reactivex.rxjava3.core.Single
-import io.reactivex.rxjava3.schedulers.Schedulers
-import org.signal.core.util.money.FiatMoney
-import org.thoughtcrime.securesms.badges.Badges
-import org.thoughtcrime.securesms.database.SignalDatabase
-import org.thoughtcrime.securesms.keyvalue.SignalStore
-import org.thoughtcrime.securesms.recipients.Recipient
-import org.thoughtcrime.securesms.storage.StorageSyncHelper
-import org.thoughtcrime.securesms.subscription.Subscription
-import org.thoughtcrime.securesms.util.PlatformCurrencyUtil
-import org.whispersystems.signalservice.api.services.DonationsService
-import org.whispersystems.signalservice.api.subscriptions.ActiveSubscription
-import org.whispersystems.signalservice.api.subscriptions.SubscriptionLevels
-import org.whispersystems.signalservice.internal.ServiceResponse
-import java.util.Currency
-import java.util.Locale
-
-/**
- * Repository which can query for the user's active subscription as well as a list of available subscriptions,
- * in the currency indicated.
- */
-class SubscriptionsRepository(private val donationsService: DonationsService) {
-
-  fun getActiveSubscription(): Single<ActiveSubscription> {
-    val localSubscription = SignalStore.donationsValues().getSubscriber()
-    return if (localSubscription != null) {
-      Single.fromCallable { donationsService.getSubscription(localSubscription.subscriberId) }
-        .subscribeOn(Schedulers.io())
-        .flatMap(ServiceResponse<ActiveSubscription>::flattenResult)
-    } else {
-      Single.just(ActiveSubscription.EMPTY)
-    }
-  }
-
-  fun getSubscriptions(): Single<List<Subscription>> = Single
-    .fromCallable { donationsService.getSubscriptionLevels(Locale.getDefault()) }
-    .subscribeOn(Schedulers.io())
-    .flatMap(ServiceResponse<SubscriptionLevels>::flattenResult)
-    .map { subscriptionLevels ->
-      subscriptionLevels.levels.map { (code, level) ->
-        Subscription(
-          id = code,
-          name = level.name,
-          badge = Badges.fromServiceBadge(level.badge),
-          prices = level.currencies.filter {
-            PlatformCurrencyUtil
-              .getAvailableCurrencyCodes()
-              .contains(it.key)
-          }.map { (currencyCode, price) ->
-            FiatMoney(price, Currency.getInstance(currencyCode))
-          }.toSet(),
-          level = code.toInt()
-        )
-      }.sortedBy {
-        it.level
-      }
-    }
-
-  fun syncAccountRecord(): Completable {
-    return Completable.fromAction {
-      SignalDatabase.recipients.markNeedsSync(Recipient.self().id)
-      StorageSyncHelper.scheduleSyncForDataChange()
-    }.subscribeOn(Schedulers.io())
-  }
-}
