+++ app/src/main/java/org/thoughtcrime/securesms/components/settings/app/subscription/donate/DonationProcessorActionResult.kt
@@ -1,12 +1,12 @@
-package org.thoughtcrime.securesms.components.settings.app.subscription.donate.stripe
+package org.thoughtcrime.securesms.components.settings.app.subscription.donate
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import org.thoughtcrime.securesms.components.settings.app.subscription.donate.gateway.GatewayRequest
 
 @Parcelize
-class StripeActionResult(
-  val action: StripeAction,
+class DonationProcessorActionResult(
+  val action: DonationProcessorAction,
   val request: GatewayRequest,
   val status: Status
 ) : Parcelable {
