+++ app/src/main/res/layout/dsl_button_primary.xml
@@ -8,9 +8,11 @@
     android:layout_height="wrap_content"
     android:layout_marginStart="@dimen/dsl_settings_gutter"
     android:layout_marginEnd="@dimen/dsl_settings_gutter"
+    android:insetTop="2dp"
+    android:insetBottom="2dp"
     app:iconGravity="textStart"
+    app:iconSize="32dp"
     app:iconTint="@null"
     tools:icon="@drawable/credit_card"
-    app:iconSize="32dp"
     tools:text="Primary button"
     tools:viewBindingIgnore="true" />
