+++ libsignal/service/src/main/java/org/whispersystems/signalservice/api/subscriptions/StripeClientSecret.java
@@ -3,13 +3,13 @@ package org.whispersystems.signalservice.api.subscriptions;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public final class SubscriptionClientSecret {
+public final class StripeClientSecret {
 
   private final String id;
   private final String clientSecret;
 
   @JsonCreator
-  public SubscriptionClientSecret(@JsonProperty("clientSecret") String clientSecret) {
+  public StripeClientSecret(@JsonProperty("clientSecret") String clientSecret) {
     this.id           = clientSecret.replaceFirst("_secret.*", "");
     this.clientSecret = clientSecret;
   }
